Add files via upload
This mod aims to overhaul the Weapon SFX of 7D2D to higher quality weapon sounds. This mod's end goal is to completely overhaul every weapon sound from melee to grenade pins being pulled. With the end, the objective is to create a more immersive world of combat for both single-player and multiplayer.

Now With Darkness Falls Compatibility